requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==3.6.4.0
blinker==1.7.0
celery==5.2.7
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
country-converter==1.2
Django==3.2.23
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
Flask==3.0.0
flower==1.2.0
geonamescache==2.0.0
h11==0.14.0
humanize==4.9.0
idna==3.6
inflection==0.5.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
json-logger==0.0.1
kombu==5.3.5
Levenshtein==0.23.0
MarkupSafe==2.1.3
numpy==1.22.4
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-Levenshtein==0.23.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
redis==5.0.1
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tornado==6.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==3.0.1
wsproto==1.2.0
